+++ test/Utils.hs
@@ -13,6 +13,7 @@ module Utils
   , shouldBeSuccess
   , shouldBeSuccessInfix
   , shouldBeSuccessOutput
+  , shouldBeSuccessOutputWithErr
   , shouldBeEmptySuccess
   , spago
   , withCwd
@@ -27,6 +28,7 @@ import qualified System.Process     as Process
 import           Test.Hspec         (HasCallStack, shouldBe, shouldSatisfy)
 import           Turtle             (ExitCode (..), FilePath, Text, cd, empty, encodeString, inproc,
                                      limit, procStrictWithErr, pwd, readTextFile, strict)
+import Data.Char (isControl)
 
 withCwd :: FilePath -> IO () -> IO ()
 withCwd dir cmd = do
@@ -62,6 +64,14 @@ shouldBeSuccessOutput expected (code, stdout, _stderr) = do
   expectedStdout <- readFixture expected
   code `shouldBe` ExitSuccess
   stdout `shouldBe` expectedStdout
+
+shouldBeSuccessOutputWithErr :: HasCallStack => FilePath -> FilePath -> (ExitCode, Text, Text) -> IO ()
+shouldBeSuccessOutputWithErr expected expectedErr (code, stdout, stderr) = do
+  expectedStdout <- readFixture expected
+  expectedStderr <- readFixture expectedErr
+  code `shouldBe` ExitSuccess
+  stdout `shouldBe` expectedStdout
+  stderr `shouldBe` expectedStderr
 
 shouldBeSuccessInfix :: HasCallStack => Text -> (ExitCode, Text, Text) -> IO ()
 shouldBeSuccessInfix expected (code, stdout, _stderr) = do